+++ src/Components/drills/ProgressBar.test.js
@@ -1,17 +1,18 @@
 import React from 'react';
 import { render, cleanup, fireEvent } from 'react-native-testing-library';
-import ProgressBar from './ProgressBar2';
+
+import ProgressBar from './ProgressBar';
 
 afterEach(cleanup);
 
-describe('<ProgressBar2 />', () => {
+describe('<ProgressBar />', () => {
   it('renders correctly', () => {
-    const mockFn = jest.fn();
-    const { toJSON, queryAllByTestId } = render(<ProgressBar total={6} current={2} onDotPress={mockFn} />);
+    const onDotPress = jest.fn();
+    const { toJSON, queryAllByTestId } = render(<ProgressBar total={6} current={2} onDotPress={onDotPress} />);
 
     expect(toJSON()).toMatchSnapshot();
 
     fireEvent.press(queryAllByTestId('progress-bar-dot')[3]);
-    expect(mockFn).toBeCalledWith(3);
+    expect(onDotPress).toBeCalledWith(3);
   });
 });
